Refactor ErrorMessage and AuthenticatedRoute
