+++ web/components/FollowLayout.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import BackButton from "@/components/BackButton";
-import ErrorMessage from "@/components/ErrorMessage";
+import ErrorMessage from "@/components/General/ErrorMessage";
 import LoadingSpinner from "@/components/LoadingSpinner";
 import UserListComponent from "@/components/UserListComponent";
 import { User } from "@/types";
