+++ web/components/General/ErrorMessage.tsx
@@ -1,3 +1,5 @@
+import { AlertTriangleIcon } from "../Icons";
+
 interface ErrorMessageProps {
   title?: string;
   message: string;
@@ -6,6 +8,7 @@ interface ErrorMessageProps {
   className?: string;
 }
 
+/** A reusable error message component with an optional retry button. */
 export default function ErrorMessage({
   title = "Something went wrong",
   message,
@@ -16,19 +19,7 @@ export default function ErrorMessage({
   return (
     <div className={`text-center p-6 ${className}`}>
       <div className="mx-auto flex items-center justify-center h-12 w-12 rounded-full bg-red-100 mb-4">
-        <svg
-          className="h-6 w-6 text-red-600"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-2.5L13.732 4c-.77-.833-1.964-.833-2.732 0L3.732 16.5c-.77.833.192 2.5 1.732 2.5z"
-          />
-        </svg>
+        <AlertTriangleIcon size={24} className="text-red-600" />
       </div>
       <h3 className="text-lg font-medium text-gray-900 mb-2">{title}</h3>
       <p className="text-gray-600 mb-4">{message}</p>
